Added support for using GET requests to get node data so that the results can be cached in the browser. There is no need for a POST if the json property is not set, and if the developer knows their tree data is safe they can enable JSON get support. More importantly this will also work if you the dataUrl points to an HTML block (so Google can index it) and it can be cached aggressively browser side.
I also modified the error handler to not display an error message if the error response text is empty. This is a workaround for issues in Firefox where Firefox will cancel any pending AJAX requests when a user browsers away from a page, so if the AJAX request for the EasyTree data is still executing when they click (I click very fast as time! lol) then you get an empty error dialog that pops up.